+++ R/compare_vectors.R
@@ -13,13 +13,14 @@
 compare_vectors <- function(x, y) {
     names <- c(deparse(substitute(x)), deparse(substitute(y)))
     u <-  sort(union(x, y))
-    x_only <-  setdiff(u, x)
-    y_only <-  setdiff(u, y)
+    x_only <- setdiff(u, x)
+    y_only <- setdiff(u, y)
     ix <- match(x_only, u)
     iy <- match(y_only, u)
     m <- cbind(x = u, y = u)
     m[iy, "y"] <- NA
     m[ix, "x"] <- NA
     dimnames(m)[[2]] <- names
+    dimnames(m)[[1]] <- u
     return(m)
 }